+++ src/components/DashboardPage.tsx
@@ -37,14 +37,13 @@ export default class DashboardPage extends React.Component<DashboardPageProps, {
   }
 
   render(): JSX.Element {
+    const { library } = this.props.params;
     return (
       <div className="dashboard">
         <Header />
         <div className="body">
-          <Stats store={this.context.editorStore} library={this.props.params.library}/>
-          { !this.props.params.library &&
-            <CirculationEvents store={this.context.editorStore} />
-          }
+          <Stats store={this.context.editorStore} library={library} />
+          <CirculationEvents store={this.context.editorStore} library={library} />
         </div>
         <Footer />
       </div>